Moving methods to generalization.
